% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNeighbours.R
\name{FindNeighbours}
\alias{FindNeighbours}
\title{FindNeighbours}
\usage{
FindNeighbours(...)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
This function can either return a \code{\link[SeuratObject]{Neighbor}} object
with the KNN information or a list of \code{\link[SeuratObject]{Graph}} objects with
the KNN and SNN depending on the settings of \code{return.neighbor} and
\code{compute.SNN}. When running on a \code{\link[SeuratObject]{Seurat}} object, this
returns the \code{\link[SeuratObject]{Seurat}} object with the Graphs or Neighbor objects
stored in their respective slots. Names of the Graph or Neighbor object can
be found with \code{\link[SeuratObject]{Graphs}} or \code{\link[SeuratObject]{Neighbors}}.
}
\description{
Masks Seurat FindNeighbors but with British English spelling
}
