% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protanCols.R
\name{protanCols}
\alias{protanCols}
\title{protanCols
Protanomaly is red-green colourblindness characterised by red-weak vision. This gives a panel of colours that are better than Seurat default for colourblind viz.
Usage: cols <- protanCols assigns the colours to a new thing}
\usage{
protanCols()
}
\value{
A list of colours
}
\description{
protanCols
Protanomaly is red-green colourblindness characterised by red-weak vision. This gives a panel of colours that are better than Seurat default for colourblind viz.
Usage: cols <- protanCols assigns the colours to a new thing
}
