% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subToMain.R
\name{subToMain}
\alias{subToMain}
\title{subToMain
Move idents from subclustered Seurat objects to an object containing the same cells by matching the cell names
Usage: data_obj <- subToMain(sub = subclustered_obj, data_obj)}
\usage{
subToMain(sub, main)
}
\arguments{
\item{sub}{A Seurat object subset from the main object with new cell idents to transfer}

\item{main}{A Seurat object containing the cell names present in the sub object}
}
\value{
A Seurat object with subclustered idents in the Ident slot
}
\description{
subToMain
Move idents from subclustered Seurat objects to an object containing the same cells by matching the cell names
Usage: data_obj <- subToMain(sub = subclustered_obj, data_obj)
}
