% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synPanel.R
\name{synPanel}
\alias{synPanel}
\title{synPanel
A list of genes for
Usage: feats <- synPanel(panel) assigns a list of relevant synovial genes to your object}
\usage{
synPanel(panel = NULL)
}
\arguments{
\item{panel}{The panel you want to call}
}
\value{
A list of relevant genes for synovial tissue clustering
}
\description{
synPanel
A list of genes for
Usage: feats <- synPanel(panel) assigns a list of relevant synovial genes to your object
}
