% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMAP.R
\name{aMAP}
\alias{aMAP}
\title{aMAP}
\arguments{
\item{seurat_object}{A Seurat object containing UMAP reduction (Seurat::RunUMAP() must have already been run)}

\item{a_centre_1}{X-coordinate of the plot center (default: calculated from data)}

\item{a_centre_2}{Y-coordinate of the plot center (default: calculated from data)}

\item{default_ext}{Default extension factor for all clusters (default: 2)}

\item{per_exts}{Named numeric vector of extension factors per cluster (default: default_ext (NULL))}

\item{text_offset}{Default text offset from label positions (default: 0.5)}

\item{per_text_offsets}{Named numeric vector of text offsets per cluster (default: text_offset (NULL))}

\item{...}{Additional parameters passed to Seurat::DimPlot() (All default Seurat::DimPlot() options are possible, including group.by, split.by, order, cols, pt.size)}
}
\value{
A ggplot2 object with annotated UMAP
}
\description{
Creates an annotated UMAP plot with extended labels pointing to cluster centers.
This function adds lines and text labels extending from cluster centroids to white space, originating from the centre of the plot,
with customizable extension and text offset per cluster, and UMAP centre parameters.
}
\examples{
# EXAMPLE 1: Cluster ordering (for plotting order on UMAP, passed to Seurat::DimPlot())
clus_order <- rev(c(
  "NK",
  "CD4 T",
  "CD8 T",
  "Cycling T",
  "B",
  "Macrophage",
  "Monocyte",
  "DC",
  "Endothelial",
  "Lining Fibroblast",
  "Sublining Fibroblast",
  "Plasma",
  "Mural"
))
# Basic usage:
# p1 <- aMAP(seurat_object = data, order = clus_order)
# p1

# EXAMPLE 2: Custom extension factors per cluster (Not all clusters have to be defined)
# (Controls how far label lines extend from cluster centers)
per_exts <- c(
  "Macrophage" = 4,
  "Cycling T" = 1.75,
  "Monocyte" = 2.5,
  "Plasma" = 2.75,
  "B" = 1.5,
  "CD4 T" = 3.5,
  "CD8 T" = 4,
  "NK" = 2.5,
  "Endothelial" = 2.5,
  "Mural" = 1.5,
  "Sublining Fibroblast" = 4.25,
  "Lining Fibroblast" = 3.75,
  "DC" = 2
)
# With custom extension factors:
# p2 <- aMAP(seurat_object = data,
#            per_exts = per_exts)
# p2

# EXAMPLE 3: Full customization with text offsets (Not all clusters have to be defined)
# (Controls distance from label line to text)
text_offsets <- c(
  "Macrophage" = 1.5,
  "Monocyte" = 2,
  "Plasma" = 1,
  "Endothelial" = 2.25
)
# Complete customization:
# p3 <- aMAP(seurat_object = data,
#            per_exts = per_exts,
#            per_text_offsets = text_offsets)
# p3

}
