% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgHeatmap.R
\name{avgHeatmap}
\alias{avgHeatmap}
\title{avgHeatmap
Plots an averaged heatmap, different from the default Seurat DoHeatmap
Usage: this will show the heatmap, use ggplot2 to save this out}
\usage{
avgHeatmap(seurat_obj, features = NULL, group.by = NULL, assay = assay)
}
\arguments{
\item{seurat_obj}{A Seurat object to plot features for}

\item{features}{A list of features to plot}

\item{group.by}{A metadata column by which to split your data}

\item{assay}{The assay in the Seurat object to pull from}
}
\value{
A ggplot2 scaled data heatmap
}
\description{
avgHeatmap
Plots an averaged heatmap, different from the default Seurat DoHeatmap
Usage: this will show the heatmap, use ggplot2 to save this out
}
