% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqCharts.R
\name{freqCharts}
\alias{freqCharts}
\title{freqCharts
Plot charts of any metadata column in a Seurat object on a barchart, both by raw numbers and by frequency (1 normalised)
Usage: freqCharts(seurat_obj, meta.1 = group, meta.2 = ident) to show split by ident per grouping, meta.1 will be X axis, meta.2 will be Y}
\usage{
freqCharts(seurat_obj, meta.1 = NULL, meta.2 = NULL, cols = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object to plot abundance charts for}

\item{meta.1}{First column in the metadata frame to use for X axis plotting, defaults to group column}

\item{meta.2}{Second column in the metadata frame to use for Y axis plotting, defaults to active idents column}

\item{cols}{Colours to use to label meta.1}
}
\value{
2 ggplot2 chart objects showing raw numbers and scaled numbers
}
\description{
freqCharts
Plot charts of any metadata column in a Seurat object on a barchart, both by raw numbers and by frequency (1 normalised)
Usage: freqCharts(seurat_obj, meta.1 = group, meta.2 = ident) to show split by ident per grouping, meta.1 will be X axis, meta.2 will be Y
}
