% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lUMAP.R
\name{lUMAP}
\alias{lUMAP}
\title{lUMAP
Different plotting for a UMAP, with points with an outline because it looks better
Usage: Makes UMAP which can be displayed by calling p, or saved with ggplot2}
\usage{
lUMAP(seurat_obj, cols = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object to plot UMAP}

\item{cols}{A list of colours to pass to the UMAP, defaults to Seurat base colours if not specified}
}
\value{
A ggplot2 UMAP
}
\description{
lUMAP
Different plotting for a UMAP, with points with an outline because it looks better
Usage: Makes UMAP which can be displayed by calling p, or saved with ggplot2
}
