% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protanCols.R
\name{protanCols}
\alias{protanCols}
\title{protanCols}
\usage{
protanCols(panel = NULL)
}
\arguments{
\item{panel}{Panel to return, one of NULL (blank) "default" or "0", "1" or "2"}
}
\value{
A list of colours
}
\description{
Protanomaly is red-green colourblindness characterised by red-weak vision. This gives a panel of colours that are better than Seurat default for colourblind viz.
Usage: cols <- protanCols assigns the colours to a new thing
}
