% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topX.R
\name{topX}
\alias{topX}
\title{A simple solution to grab top X genes per cluster from a FindAllMarkers readout.}
\usage{
topX(res, n_genes = 5, pct_cut = 0.1, p_cut = 0.05, cut_by = "pct.1", neg = F)
}
\arguments{
\item{res}{The FindAllMarkers readout that you want to select from.}

\item{n_genes}{Number of top genes per cluster to return. Default is 5.}

\item{pct_cut}{Minimum percent expression cutoff. Default is 0.1 (10\%).}

\item{cut_by}{A column in the results dataframe by which to cut. Default is pct.1.}

\item{p_cut}{Adjusted p-value cutoff. Default is 0.05.}

\item{neg}{Binary which determines whether you want negative markers. Default is False.}
}
\value{
A tibble of top X genes for each cluster in your readout
}
\description{
A simple solution to grab top X genes per cluster from a FindAllMarkers readout.
}
